move shared EntityStatuses enum to types
