+++ services/ui-src/src/components/index.ts
@@ -110,7 +110,7 @@ export { StatusTable } from "./statusing/StatusTable";
 // tables
 export { EntityRow } from "./tables/EntityRow";
 export { Table } from "./tables/Table";
-export { EntityStatusIcon, EntityStatuses } from "./tables/EntityStatusIcon";
+export { EntityStatusIcon } from "./tables/EntityStatusIcon";
 // widgets
 export { SpreadsheetWidget } from "./widgets/SpreadsheetWidget";
 // Redirects
