+++ services/ui-src/src/types/other.ts
@@ -64,6 +64,13 @@ export enum PageTypes {
   REVIEW_SUBMIT = "reviewSubmit",
 }
 
+export enum EntityStatuses {
+  COMPLETE = "complete",
+  CLOSE = "close",
+  NO_STATUS = "no status",
+  DISABLED = "disabled",
+}
+
 export interface InputChangeEvent extends React.ChangeEvent<HTMLInputElement> {}
 
 export type { IconType } from "react-icons";
